Extract Params component
fixes #127
